+++ CHANGELOG.md
@@ -1,3 +1,27 @@
+## v1.2.0
+- Added UI to the freecam view
+- Added an option to automatically reset the freecam transform when activating freecam
+- Added an option to automatically reset the freecam transform when activating freecam while far away from the freecam
+- Fixed an issue where typing in the terminal could cause the freecam to activate
+- Fixed an issue where typing in the chat could cause the freecam to activate
+- Fixed an issue where the player could toggle crouch when in freecam and having player movement set to off
+- Fixed an issue where the player could jump when in freecam and having player movement set to off
+- Fixed an issue where the player could interact when in freecam and having player movement set to off
+- Fixed an issue where the player could use items when in freecam and having player movement set to off
+- Fixed an issue where the player could switch items when in freecam and having player movement set to off
+- Fixed an issue where the player could emote when in freecam and having player movement set to off
+- Fixed an issue where the player could inspect when in freecam and having player movement set to off
+- Fixed an issue where the player could scan when in freecam and having player movement set to off
+- Fixed an issue where the player could boost in a vehicle when in freecam and having player movement set to off
+- Fixed an issue where the freecam would glitch when using the terminal while in freecam
+- Fixed an issue where the freecam would spawn at an inconsistent height behind the player
+- Fixed an issue where the freecam would not detect the player as in a vehicle when in a passenger seat
+- Fixed an issue where the interact UI would shown up when in freecam
+- Fixed an issue where the scanned items UI would shown up when in freecam
+- Fixed an issue where the freecam sometimes would remain stuck in place when using freecam while in a vehicle
+- Fixed an issue where the freecam speed would be different when using freecam while in a vehicle
+- Fixed an issue where the freecam speed/fov scrolling would increment 4 times as much as intended
+
 ## v1.1.1
 - Fixed readme formatting